norme fixes
