+++ Makefile
@@ -145,6 +145,7 @@ LINE_EDITION_FILE += home_end_key.c
 LINE_EDITION_FILE += left_right_bs_key.c
 LINE_EDITION_FILE += read_user_input.c
 LINE_EDITION_FILE += reverse_search.c
+LINE_EDITION_FILE += reverse_search_row_count.c
 LINE_EDITION_FILE += reverse_search_utils.c
 LINE_EDITION_FILE += row_count.c
 LINE_EDITION_FILE += term_attr.c
